Refactor: Cars 생성 테스트를 CarsTest로 이동
